separete params and entities types
